Ehr-10200 merge-scheduler-work (#722)
All of the scheduler related work have been merged in this one. Code
review has passed for those branches:
https://github.com/Psychplus/ReactUI/pull/633
https://github.com/Psychplus/ReactUI/pull/650
https://github.com/Psychplus/ReactUI/pull/710

P.S.
This PR also includes all the qa feedback changes

---------

+++ apps/galaxy/src/ui/visit/edit-visit/components/staff-comments/treatment-tab/action-cell.tsx
@@ -5,8 +5,11 @@ import { IconButton } from '@radix-ui/themes'
 import { Trash2 } from 'lucide-react'
 import toast from 'react-hot-toast'
 import { PropsWithRow } from '@/components'
+import { useHasPermission } from '@/hooks'
 import { STAFF_COMMENT_STATUS, StaffComment } from '@/types'
+import { PermissionAlert } from '@/ui/schedule/shared'
 import { deleteStaffCommentAction } from '@/ui/staff-comments/actions'
+import { ADD_DELETE_STAFF_COMMENT_EDIT_VISIT_POPUP } from '@/ui/visit/constants'
 import { StaffCommentParams } from '@/ui/visit/types'
 
 const ActionCell = ({
@@ -17,14 +20,20 @@ const ActionCell = ({
 }: PropsWithRow<StaffComment> & {
   fetchComments: (payload: StaffCommentParams) => void
 }) => {
+  const [isOpen, setIsOpen] = useState<boolean>(false)
   const [loading, setLoading] = useState<boolean>(false)
-
+  const canAddDeleteStaffComment = useHasPermission(
+    'staffCommentAddDeleteEditVisitPopup',
+  )
   const handleDelete = async () => {
+    if (!canAddDeleteStaffComment) {
+      return setIsOpen(true)
+    }
     setLoading(true)
 
     const result = await deleteStaffCommentAction(id)
     if (result.state === 'error') {
-      toast.error(result.error)
+      toast.error(result.error || 'Failed to delete comment')
       return
     }
 
@@ -39,15 +48,22 @@ const ActionCell = ({
     setLoading(false)
   }
   return (
-    <IconButton
-      variant="ghost"
-      color="gray"
-      size="1"
-      disabled={loading}
-      onClick={handleDelete}
-    >
-      <Trash2 size={16} />
-    </IconButton>
+    <>
+      <PermissionAlert
+        isOpen={isOpen}
+        onClose={() => setIsOpen(false)}
+        message={ADD_DELETE_STAFF_COMMENT_EDIT_VISIT_POPUP}
+      />
+      <IconButton
+        variant="ghost"
+        color="gray"
+        size="1"
+        disabled={loading}
+        onClick={handleDelete}
+      >
+        <Trash2 size={16} />
+      </IconButton>
+    </>
   )
 }
